임동준_20240621_숙소리스트, 상세보기(숙소 등록 진행중)

+++ src/main/java/com/lec/spring/domain/ProvLodging.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProvLodging {
-    private int lodgingId;                 // 숙소 ID
+    private Long lodgingId;                 // 숙소 ID
     private String lodgingPicture1;        // 숙소 사진 1
     private String lodgingPicture2;        // 숙소 사진 2
     private String lodgingPicture3;        // 숙소 사진 3
